+++ stepup/reprep/convert_markdown.py
@@ -22,91 +22,40 @@
 import argparse
 import re
 import sys
-import tempfile
 
 import markdown
+
+from .render import render
 
 __all__ = ("convert_markdown",)
 
 
-HTML_HEADER = """\
+HTML_TEMPLATE = """\
 <!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" lang="" xml:lang="">
 <head>
   <meta charset="utf-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=yes" />
-  <title>Preview</title>
-  <style>
-    html {
-      line-height: 1.3;
-      font-family: IBM Plex Sans, Georgia, serif;
-      font-size: 10pt;
-    }
-    h1, h2 {
-      margin-top: 1cm;
-      border-bottom: 1pt solid #DDDDDD;
-    }
-    h3, h4, p {
-      margin-top: 0.5cm;
-      padding-left: 1.0cm;
-    }
-    ul {
-      padding-left: 2.0cm;
-    }
-    ul ul {
-      padding-left: 1.0cm;
-    }
-    li {
-      margin-top: 2pt;
-      margin-bottom: 2pt;
-    }
-    code {
-      font-family: IBM Plex Mono;
-      background-color: rgba(175, 184, 193, 0.2);
-      padding: 2pt 5pt 2pt 5pt;
-      border-radius: 6pt;
-    }
-    table {
-      margin-left: auto;
-      margin-right: auto;
-      text-align: center;
-      border-collapse: collapse;
-    }
-    td, th {
-      border: 1pt solid #BBBBBB;
-      padding: 4pt;
-      font-size: 10pt;
-    }
-    @page {
-        size: A4;
-        margin: 1.5cm 1.5cm 1.5cm 1.5cm;
-    }
-  </style>
+  <title>{{ title }}</title>
+  {{ css | indent(width=2) }}
 </head>
 <body>
+  {{ body | indent(width=2) }}
+</body>
+</html>
 """
 
 
-HTML_FOOTER = "</body>"
-
-
-MACRO_TEXT = r"""\
-\bvec:\vec{\mathbf{#1}}
-\normvec:\hat{\mathbf{#1}}
-\d:\operatorname{d}\!{#1}
-\ihat:\hat{\mathbf{i}}
-\jhat:\hat{\mathbf{j}}
-\khat:\hat{\mathbf{k}}
-"""
-
-
-def convert_markdown(text_md: str) -> str:
+def convert_markdown(text_md: str, katex: bool = False, path_macro: str | None = None) -> str:
     """Convert Markdown to HTML with KaTeX support.
 
     Parameters
     ----------
     text_md
-        The markdown source text
+        The markdown source text.
+    katex
+        Set to `True` to enable KaTeX support.
+    path_macro
+        A file with KaTeX macro definitions.
 
     Returns
     -------
@@ -116,27 +65,34 @@ def convert_markdown(text_md: str) -> str:
     # Convert conventional LaTeX equation syntax to make it compatible with markdown_katex
     text_md = re.sub(r"\B\$(\S|\S[^\n\r]*?\S)\$\B", r"$`\1`$", text_md)
 
-    # Write macros to temporary file for KaTeX.
-    with tempfile.NamedTemporaryFile(suffix=".tex") as f:
-        f.write(MACRO_TEXT.encode("ascii"))
-        f.flush()
-        fn_macro = f.name
+    extensions = [
+        "fenced_code",
+        "tables",
+        "meta",
+    ]
+    configs = {}
 
-        md_ctx = markdown.Markdown(
-            extensions=[
-                "fenced_code",
-                "markdown_katex",
-                "tables",
-            ],
-            extension_configs={
-                "markdown_katex": {"insert_fonts_css": True, "macro-file": fn_macro}
-            },
-        )
+    if katex:
+        extensions.append("markdown_katex")
+        configs["markdown_katex"] = {"insert_fonts_css": True}
+        if path_macro is not None:
+            configs["markdown_katex"]["macro-file"] = path_macro
 
-        # Convert to HTML
-        text_html = HTML_HEADER + md_ctx.convert(text_md) + HTML_FOOTER
+    md_ctx = markdown.Markdown(
+        extensions=extensions,
+        extension_configs=configs,
+    )
 
-    return text_html
+    # Convert to HTML
+    body = md_ctx.convert(text_md)
+    variables = {
+        "body": body,
+        "title": md_ctx.Meta.get("title", ["Untitled"])[0],
+        "css": "\n".join(
+            f'<link rel="stylesheet" href="{path_css}">' for path_css in md_ctx.Meta.get("css", [])
+        ),
+    }
+    return render("HTML_TEMPLATE", variables, str_in=HTML_TEMPLATE)
 
 
 def parse_args() -> argparse.Namespace:
@@ -146,6 +102,8 @@ def parse_args() -> argparse.Namespace:
     )
     parser.add_argument("markdown", help="A Markdown file with extension `.md`")
     parser.add_argument("html", help="A HTML output filename")
+    parser.add_argument("--katex", default=False, action="store_true", help="Enable KaTeX")
+    parser.add_argument("--katex-macros", default=None, help="KaTeX macro file")
     return parser.parse_args()
 
 
@@ -155,7 +113,7 @@ def main() -> int:
     if not args.markdown.endswith(".md"):
         raise ValueError("The markdown file must end with the .md extension.")
     with open(args.markdown) as fm, open(args.html, "w") as fh:
-        fh.write(convert_markdown(fm.read()))
+        fh.write(convert_markdown(fm.read(), args.katex, args.katex_macros))
     return 0